Split out confirmDelete
